Race improvement
Race class was updated to accept any number of horses if the value does not exceed the maximum chosen value (currently 15).
Additionally, security functions were added to prevent errors occurring when the value for number of horses or length of the race is incorrect. Also, infinite loop bug was resolved. It occurred when all horses failed before any of them reached the finish line. Now the race ends immediately, if all horses fell.
Also main file was updated to be able to add any number of horses to the race by the choice of the user. Randomly chosen symbols are assigned to each horse. Also minimum confidence generated was increased from 0 to 0.1, as too many incredibly slow horses were created during testing.