More information output and main() update
Race class was updated and now outputs the name and current confidence of each horse during the race. Additionally, when the race finishes, an appropriate message is printed out. According to the result of the race the confidence of each horse is adjusted.
HorceRacing.main() was updated to give the user an option to start another round with the same settings after every race.