Update Horse Class
Implemented all fields and methods.
File Rearranging and main file edits.
Moved all the files for the textual version of horse racing to Part1 folder and created HorseRacing.java file. At this point it just starts a race with distance chosen by the user and three horses with random confidence.